REF: Consolidate Frame/Series arithmetic helpers in ops.methods
