+++ pandas/core/frame.py
@@ -15,7 +15,6 @@ from collections import abc
 import functools
 from io import StringIO
 import itertools
-import operator
 import sys
 from textwrap import dedent
 from typing import (
@@ -89,7 +88,6 @@ from pandas.core.dtypes.cast import (
 from pandas.core.dtypes.common import (
     infer_dtype_from_object,
     is_1d_only_ea_dtype,
-    is_array_like,
     is_bool_dtype,
     is_dataclass,
     is_dict_like,
@@ -117,8 +115,6 @@ from pandas.core import (
     algorithms,
     common as com,
     nanops,
-    ops,
-    roperator,
 )
 from pandas.core.accessor import CachedAccessor
 from pandas.core.apply import (
@@ -492,7 +488,8 @@ ValueError: columns overlap but no suffix specified:
 # DataFrame class
 
 
-class DataFrame(FrameOps, NDFrame, OpsMixin):
+# Cannot override writable attribute "_get_axis_number" with a final one
+class DataFrame(NDFrame, FrameOps, OpsMixin):  # type: ignore[misc]
     """
     Two-dimensional, size-mutable, potentially heterogeneous tabular data.
 
@@ -7450,386 +7447,6 @@ class DataFrame(FrameOps, NDFrame, OpsMixin):
             assert isinstance(result.columns, MultiIndex)
             result.columns = result.columns.reorder_levels(order)
         return result
-
-    # ----------------------------------------------------------------------
-    # Arithmetic Methods
-
-    def _cmp_method(self, other, op):
-        axis: Literal[1] = 1  # only relevant for Series other case
-
-        self, other = self._align_for_op(other, axis, flex=False, level=None)
-
-        # See GH#4537 for discussion of scalar op behavior
-        new_data = self._dispatch_frame_op(other, op, axis=axis)
-        return self._construct_result(new_data)
-
-    def _arith_method(self, other, op):
-        if self._should_reindex_frame_op(other, op, 1, None, None):
-            return self._arith_method_with_reindex(other, op)
-
-        axis: Literal[1] = 1  # only relevant for Series other case
-        other = ops.maybe_prepare_scalar_for_op(other, (self.shape[axis],))
-
-        self, other = self._align_for_op(other, axis, flex=True, level=None)
-
-        new_data = self._dispatch_frame_op(other, op, axis=axis)
-        return self._construct_result(new_data)
-
-    _logical_method = _arith_method
-
-    def _arith_method_with_reindex(self, right: DataFrame, op) -> DataFrame:
-        """
-        For DataFrame-with-DataFrame operations that require reindexing,
-        operate only on shared columns, then reindex.
-
-        Parameters
-        ----------
-        right : DataFrame
-        op : binary operator
-
-        Returns
-        -------
-        DataFrame
-        """
-        left = self
-
-        # GH#31623, only operate on shared columns
-        cols, lcols, rcols = left.columns.join(
-            right.columns, how="inner", level=None, return_indexers=True
-        )
-
-        new_left = left.iloc[:, lcols]
-        new_right = right.iloc[:, rcols]
-        result = op(new_left, new_right)
-
-        # Do the join on the columns instead of using left._align_for_op
-        #  to avoid constructing two potentially large/sparse DataFrames
-        join_columns, _, _ = left.columns.join(
-            right.columns, how="outer", level=None, return_indexers=True
-        )
-
-        if result.columns.has_duplicates:
-            # Avoid reindexing with a duplicate axis.
-            # https://github.com/pandas-dev/pandas/issues/35194
-            indexer, _ = result.columns.get_indexer_non_unique(join_columns)
-            indexer = algorithms.unique1d(indexer)
-            result = result._reindex_with_indexers(
-                {1: [join_columns, indexer]}, allow_dups=True
-            )
-        else:
-            result = result.reindex(join_columns, axis=1)
-
-        return result
-
-    def _should_reindex_frame_op(
-        self: DataFrame, right, op, axis: int, fill_value, level
-    ) -> bool:
-        """
-        Check if this is an operation between DataFrames that will need to reindex.
-        """
-        if op is operator.pow or op is roperator.rpow:
-            # GH#32685 pow has special semantics for operating with null values
-            return False
-
-        if not isinstance(right, DataFrame):
-            return False
-
-        if fill_value is None and level is None and axis == 1:
-            # TODO: any other cases we should handle here?
-
-            # Intersection is always unique so we have to check the unique columns
-            left_uniques = self.columns.unique()
-            right_uniques = right.columns.unique()
-            cols = left_uniques.intersection(right_uniques)
-            if len(cols) and not (
-                len(cols) == len(left_uniques) and len(cols) == len(right_uniques)
-            ):
-                # TODO: is there a shortcut available when len(cols) == 0?
-                return True
-
-        return False
-
-    def _align_for_op(
-        self, other, axis, flex: bool | None = False, level: Level = None
-    ):
-        """
-        Convert rhs to meet lhs dims if input is list, tuple or np.ndarray.
-
-        Parameters
-        ----------
-        left : DataFrame
-        right : Any
-        axis : int, str, or None
-        flex : bool or None, default False
-            Whether this is a flex op, in which case we reindex.
-            None indicates not to check for alignment.
-        level : int or level name, default None
-
-        Returns
-        -------
-        left : DataFrame
-        right : Any
-        """
-        left, right = self, other
-
-        def to_series(right):
-            msg = (
-                "Unable to coerce to Series, "
-                "length must be {req_len}: given {given_len}"
-            )
-
-            # pass dtype to avoid doing inference, which would break consistency
-            #  with Index/Series ops
-            dtype = None
-            if getattr(right, "dtype", None) == object:
-                # can't pass right.dtype unconditionally as that would break on e.g.
-                #  datetime64[h] ndarray
-                dtype = object
-
-            if axis is not None and left._get_axis_name(axis) == "index":
-                if len(left.index) != len(right):
-                    raise ValueError(
-                        msg.format(req_len=len(left.index), given_len=len(right))
-                    )
-                right = left._constructor_sliced(right, index=left.index, dtype=dtype)
-            else:
-                if len(left.columns) != len(right):
-                    raise ValueError(
-                        msg.format(req_len=len(left.columns), given_len=len(right))
-                    )
-                right = left._constructor_sliced(right, index=left.columns, dtype=dtype)
-            return right
-
-        if isinstance(right, np.ndarray):
-            if right.ndim == 1:
-                right = to_series(right)
-
-            elif right.ndim == 2:
-                # We need to pass dtype=right.dtype to retain object dtype
-                #  otherwise we lose consistency with Index and array ops
-                dtype = None
-                if getattr(right, "dtype", None) == object:
-                    # can't pass right.dtype unconditionally as that would break on e.g.
-                    #  datetime64[h] ndarray
-                    dtype = object
-
-                if right.shape == left.shape:
-                    right = left._constructor(
-                        right, index=left.index, columns=left.columns, dtype=dtype
-                    )
-
-                elif right.shape[0] == left.shape[0] and right.shape[1] == 1:
-                    # Broadcast across columns
-                    right = np.broadcast_to(right, left.shape)
-                    right = left._constructor(
-                        right, index=left.index, columns=left.columns, dtype=dtype
-                    )
-
-                elif right.shape[1] == left.shape[1] and right.shape[0] == 1:
-                    # Broadcast along rows
-                    right = to_series(right[0, :])
-
-                else:
-                    raise ValueError(
-                        "Unable to coerce to DataFrame, shape "
-                        f"must be {left.shape}: given {right.shape}"
-                    )
-
-            elif right.ndim > 2:
-                raise ValueError(
-                    "Unable to coerce to Series/DataFrame, "
-                    f"dimension must be <= 2: {right.shape}"
-                )
-
-        elif is_list_like(right) and not isinstance(right, (Series, DataFrame)):
-            # GH#36702. Raise when attempting arithmetic with list of array-like.
-            if any(is_array_like(el) for el in right):
-                raise ValueError(
-                    f"Unable to coerce list of {type(right[0])} to Series/DataFrame"
-                )
-            # GH#17901
-            right = to_series(right)
-
-        if flex is not None and isinstance(right, DataFrame):
-            if not left._indexed_same(right):
-                if flex:
-                    left, right = left.align(
-                        right, join="outer", level=level, copy=False
-                    )
-                else:
-                    raise ValueError(
-                        "Can only compare identically-labeled (both index and columns) "
-                        "DataFrame objects"
-                    )
-        elif isinstance(right, Series):
-            # axis=1 is default for DataFrame-with-Series op
-            axis = left._get_axis_number(axis) if axis is not None else 1
-
-            if not flex:
-                if not left.axes[axis].equals(right.index):
-                    raise ValueError(
-                        "Operands are not aligned. Do "
-                        "`left, right = left.align(right, axis=1, copy=False)` "
-                        "before operating."
-                    )
-
-            left, right = left.align(
-                right, join="outer", axis=axis, level=level, copy=False
-            )
-            right = left._maybe_align_series_as_frame(right, axis)
-
-        return left, right
-
-    def _maybe_align_series_as_frame(self, series: Series, axis: AxisInt):
-        """
-        If the Series operand is not EA-dtype, we can broadcast to 2D and operate
-        blockwise.
-        """
-        rvalues = series._values
-        if not isinstance(rvalues, np.ndarray):
-            # TODO(EA2D): no need to special-case with 2D EAs
-            if rvalues.dtype in ("datetime64[ns]", "timedelta64[ns]"):
-                # We can losslessly+cheaply cast to ndarray
-                rvalues = np.asarray(rvalues)
-            else:
-                return series
-
-        if axis == 0:
-            rvalues = rvalues.reshape(-1, 1)
-        else:
-            rvalues = rvalues.reshape(1, -1)
-
-        rvalues = np.broadcast_to(rvalues, self.shape)
-        # pass dtype to avoid doing self
-        return type(self)(
-            rvalues, index=self.index, columns=self.columns, dtype=rvalues.dtype
-        )
-
-    def _dispatch_frame_op(self, right, func: Callable, axis: AxisInt | None = None):
-        """
-        Evaluate the frame operation func(left, right) by evaluating
-        column-by-column, dispatching to the Series implementation.
-
-        Parameters
-        ----------
-        right : scalar, Series, or DataFrame
-        func : arithmetic or comparison operator
-        axis : {None, 0, 1}
-
-        Returns
-        -------
-        DataFrame
-        """
-        # Get the appropriate array-op to apply to each column/block's values.
-        array_op = ops.get_array_op(func)
-
-        right = lib.item_from_zerodim(right)
-        if not is_list_like(right):
-            # i.e. scalar, faster than checking np.ndim(right) == 0
-            with np.errstate(all="ignore"):
-                bm = self._mgr.apply(array_op, right=right)
-            return self._constructor(bm)
-
-        elif isinstance(right, DataFrame):
-            assert self.index.equals(right.index)
-            assert self.columns.equals(right.columns)
-            # TODO: The previous assertion `assert right._indexed_same(self)`
-            #  fails in cases with empty columns reached via
-            #  _arith_method_with_reindex
-
-            # TODO operate_blockwise expects a manager of the same type
-            with np.errstate(all="ignore"):
-                bm = self._mgr.operate_blockwise(
-                    # error: Argument 1 to "operate_blockwise" of "ArrayManager" has
-                    # incompatible type "Union[ArrayManager, BlockManager]"; expected
-                    # "ArrayManager"
-                    # error: Argument 1 to "operate_blockwise" of "BlockManager" has
-                    # incompatible type "Union[ArrayManager, BlockManager]"; expected
-                    # "BlockManager"
-                    right._mgr,  # type: ignore[arg-type]
-                    array_op,
-                )
-            return self._constructor(bm)
-
-        elif isinstance(right, Series) and axis == 1:
-            # axis=1 means we want to operate row-by-row
-            assert right.index.equals(self.columns)
-
-            right = right._values
-            # maybe_align_as_frame ensures we do not have an ndarray here
-            assert not isinstance(right, np.ndarray)
-
-            with np.errstate(all="ignore"):
-                arrays = [
-                    array_op(_left, _right)
-                    for _left, _right in zip(self._iter_column_arrays(), right)
-                ]
-
-        elif isinstance(right, Series):
-            assert right.index.equals(self.index)  # Handle other cases later
-            right = right._values
-
-            with np.errstate(all="ignore"):
-                arrays = [array_op(left, right) for left in self._iter_column_arrays()]
-
-        else:
-            # Remaining cases have less-obvious dispatch rules
-            raise NotImplementedError(right)
-
-        return type(self)._from_arrays(
-            arrays, self.columns, self.index, verify_integrity=False
-        )
-
-    def _combine_frame(self, other: DataFrame, func, fill_value=None):
-        # at this point we have `self._indexed_same(other)`
-
-        if fill_value is None:
-            # since _arith_op may be called in a loop, avoid function call
-            #  overhead if possible by doing this check once
-            _arith_op = func
-
-        else:
-
-            def _arith_op(left, right):
-                # for the mixed_type case where we iterate over columns,
-                # _arith_op(left, right) is equivalent to
-                # left._binop(right, func, fill_value=fill_value)
-                left, right = ops.fill_binop(left, right, fill_value)
-                return func(left, right)
-
-        new_data = self._dispatch_frame_op(other, _arith_op)
-        return new_data
-
-    def _construct_result(self, result) -> DataFrame:
-        """
-        Wrap the result of an arithmetic, comparison, or logical operation.
-
-        Parameters
-        ----------
-        result : DataFrame
-
-        Returns
-        -------
-        DataFrame
-        """
-        out = self._constructor(result, copy=False).__finalize__(self)
-        # Pin columns instead of passing to constructor for compat with
-        #  non-unique columns case
-        out.columns = self.columns
-        out.index = self.index
-        return out
-
-    def __divmod__(self, other) -> tuple[DataFrame, DataFrame]:
-        # Naive implementation, room for optimization
-        div = self // other
-        mod = self - div * other
-        return div, mod
-
-    def __rdivmod__(self, other) -> tuple[DataFrame, DataFrame]:
-        # Naive implementation, room for optimization
-        div = other // self
-        mod = other - div * self
-        return div, mod
 
     # ----------------------------------------------------------------------
     # Combination-Related
